pkg/redirectpolicy: Make code robust against incorrect policy configurations
If an incorrect service name is specified in an LRP, we need to guard against
this case when service IP can't be retrieved.
Also, when a valid LRP is applied before the service it selects, service information
won't be available in the manager callback to add an LRP. The LRP will be
applied when the service callback is later received by the manager.

Signed-off-by: Aditi Ghag <aditi@cilium.io>