policy: Add ICMP and ICMPv6 API with feature flag
This commit implements ICMP/ICMPv6 API for CNP and CCNP.
The `icmp` field is added to CNP and CCNP manifests, and policy can be added by setting ICMP type to that field.
Currentlly, this ICMP CNP feature is disabled by default because of the kernel complexity issue, and it is enabled when the "enable-icmp-rule" hidden flag is set.

Example manifest:
```
apiVersion: "cilium.io/v2"
kind: CiliumNetworkPolicy
metadata:
  name: "cnp-icmp"
spec:
  endpointSelector:
    matchLabels:
      org: empire
      class: deathstar
  ingress:
  - fromEndpoints:
    - matchLabels:
        org: empire
    icmps:
    - fields:
      - type: 8
      # ICMPv6
      - type: 128
        family: IPv6
```

Signed-off-by: Tomoki Sugiura <cheztomo513@gmail.com>