test/K8sBookInfo: Readiness probes for test pods
The test pods used in the k8sBookInfo test did not have readiness probe.
Therefore, from time to time, the pods would be marked as Ready before
their application was listening on the server's port. The subsequent
connectivity check would thus fail with a TCP RST.

This commit fixes that issue by defining a proper readiness probe for
all pods used in K8sBookInfo, such that the test will only proceed once
the application is ready to serve requests.

Signed-off-by: Paul Chaignon <paul@cilium.io>