policy: Fix cilium policy trace output when only deny rules are applied
Currently, deny rules are not counted in `*Repository.GetRulesMatching()`.
Therefore, this function returns `false` even though a rule is applied if it is a deny rule.
This causes a wrong output of `cilium policy trace` commnad in some cases.

This PR makes `*Repository.GetRulesMatching()` to count ingress/egress deny rules.
Also, add the unit test for this.

Signed-off-by: Tomoki Sugiura <cheztomo513@gmail.com>