bpf: Remove macros for service types
This commit removes ENABLE_LOADBALANCER, ENABLE_EXTERNAL_IP, and
ENABLE_HOSTPORT. The agent-level options are kept and, if disabled,
services of the corresponding types won't be populated in BPF maps.

The goal of this commit is to reduce the number of datapath options we
need to cover in complexity tests. Removing 3 options can lead to as
much as an x8 reduction. Since we won't populate the BPF maps if those
options are disabled, we don't really need the macros.

Signed-off-by: Paul Chaignon <paul@cilium.io>