routing: Fix incorrect interface selection for pod routes
The Configure method relies on the MAC address to select the proper
egress interface for new pods (EKS and AKS). Several interfaces can
however have the same MAC address in the case of slave devices. In
such a case, the wrong interface may be selected.

To avoid this, we skip Linux slave devices during the lookup by MAC
address. Thus, in case of slave devices, we will select the master
device.

Fixes: 26308b63 ("Implement support for cilium-health in ENI mode")
Signed-off-by: Paul Chaignon <paul@cilium.io>